refactor response models, change tollfees to ints, add globalexceptionhandler and remove try/catch in controller/applicationlayer

+++ Nb.Tolls.Domain/Results/TollFeeResult.cs
@@ -3,5 +3,5 @@
 public class TollFeeResult
 {
     public DateTime TollFeeTime { get; init; }
-    public decimal TollFee { get; init; }
+    public int TollFee { get; init; }
 }
